backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.31
alembic>=1.13.1
python-multipart>=0.0.9
pillow>=11.0.0
scikit-learn>=1.5.0
numpy>=2.0.0
rembg>=2.0.57
onnxruntime>=1.18.0
colormath>=3.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
pytest>=8.2.2
pytest-asyncio>=0.23.7
